api/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
pkg-resources==0.0.0
uWSGI==2.0.18
Werkzeug==0.15.1
mysql-connector==2.1.6